refactor: query countries in Cards component
