+++ src/App.css
@@ -1,5 +1,6 @@
 .App {
   text-align: center;
+  /* width: 100%; */
 }
 
 .App-header {
