+++ src/components/Card/Card.module.css
@@ -2,7 +2,7 @@
     padding: 1rem;
     position: relative;
     width: 25%;
-    /* max-width: 14rem; ? */
+    max-height: 14rem;
     background-color: white;
     border-radius: 10px;
     color: #282c34;
